requirements.txt
# Requirements for Fire Detection Project

# Web Demo
streamlit>=1.32.0

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# YOLOv11
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# Configuration
pyyaml>=6.0

# Progress Bar
tqdm>=4.65.0

# Optional: For GPU support
# torch-cuda (install separately based on CUDA version)

# Optional: Export formats
# onnx>=1.14.0
# onnxruntime>=1.15.0

# Optional: Video processing
# imageio>=2.31.0
# imageio-ffmpeg>=0.4.8
